% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepr.R
\name{grepr}
\alias{grepr}
\title{Search for patterns in files}
\usage{
grepr(pattern)
}
\arguments{
\item{pattern}{a character string containing a regular expression}
}
\description{
Search for patterns in files
}
