% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_umlauts.R
\name{grep_umlauts}
\alias{grep_umlauts}
\alias{print_substitution_table}
\title{Mark all umlauts}
\usage{
grep_umlauts(pattern = specials, ..., ignore_roxygen = FALSE)

print_substitution_table()
}
\arguments{
\item{pattern}{a character string containing a regular expression}

\item{...}{arguments which are passed to \code{\link[=grepr]{grepr()}}}

\item{ignore_roxygen}{do not show umlauts in roxygen blocks or Rd files}
}
\description{
Searches for umlauts (ÖÄÜöäü) with \code{\link[=grepr]{grepr()}} and creates RStudio markers
}
